[change] Moved device API view writing logic to background worker #484
Closes #484
